Aggiunte e testate le funzionalità inserimento Società e Dipendente, ricerca Società per esempio e rimozione Società

+++ src/main/java/it/prova/gestionesocieta/service/SocietaServiceImpl.java
@@ -6,9 +6,11 @@ import javax.persistence.EntityManager;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import it.prova.gestionesocieta.exception.SocietaConAncoraDipendentiException;
 import it.prova.gestionesocieta.model.Societa;
 import it.prova.gestionesocieta.repository.SocietaRepository;
 
@@ -30,6 +32,11 @@ public class SocietaServiceImpl implements SocietaService{
 		
 		return societaRepository.findById(id).orElse(null);
 	}
+	
+	@Transactional(readOnly = true)
+	public Societa caricaSingolaSocietaFetch(Long id) {
+		return societaRepository.findByIdEager(id);
+	}
 
 	@Transactional
 	public void aggiorna(Societa societa) {
@@ -43,7 +50,9 @@ public class SocietaServiceImpl implements SocietaService{
 	}
 
 	@Transactional
-	public void rimuovi(Societa societa) {
+	public void rimuovi(Societa societa) throws SocietaConAncoraDipendentiException {
+		if(!societaRepository.findByIdEager(societa.getId()).getDipendenti().isEmpty())
+			throw new SocietaConAncoraDipendentiException("Impossibile rimuovere, ci sono ancora Dipendenti di questa societa' sul db.");
 		societaRepository.delete(societa);
 	}
 
@@ -52,11 +61,11 @@ public class SocietaServiceImpl implements SocietaService{
 		String query = "select s from Societa s where s.id=s.id";
 		
 		if(StringUtils.isNotEmpty(example.getRagioneSociale()))
-			query += " and s.ragionesociale like '%"+example.getRagioneSociale()+"%'";
+			query += " and s.ragioneSociale like '%"+example.getRagioneSociale()+"%'";
 		if(StringUtils.isNotEmpty(example.getIndirizzo()))
 			query += " and s.indirizzo like '%"+example.getIndirizzo()+"%'";
 		if(example.getDataFondazione()!=null)
-			query += " and s.datacreazione > '"+example.getDataFondazione()+"'";
+			query += " and s.dataFondazione > '"+example.getDataFondazione().toInstant()+"'";
 		
 		return entityManager.createQuery(query, Societa.class).getResultList();
 	}
